Client Side

+++ js/Storage.js
@@ -156,32 +156,37 @@
 			 */
 			clear: function(key, value, callback){
 				try{
-					// clear all storage
-					if(key == null || key == undefined || key == ""){
-						localStorage.clear();
-						callback("sucsess", null);
-						return;
-					}
+					if(Globals.noLockedItemrs()){
+						// clear all storage
+						if(key == null || key == undefined || key == ""){
+							localStorage.clear();
+							callback("sucsess", null);
+							return;
+						}
 
-					// remove specific key from storage
-					if(value == null || value == undefined || value == ""){
-						localStorage.removeItem(key);
-						callback(JSON.parse(localStorage.getItem(key)), null);
-						return;
-					}
-
-					//remove specific value from key in storage
-					var tempData = JSON.parse(localStorage.getItem(key));
-					for (var i = tempData.length - 1; i >= 0; i--) {
-						if( JSON.stringify(value) == JSON.stringify(tempData[i]) ){
-							tempData.splice(i, 1);
-							localStorage.setItem(key, tempData);
+						// remove specific key from storage
+						if(value == null || value == undefined || value == ""){
+							localStorage.removeItem(key);
 							callback(JSON.parse(localStorage.getItem(key)), null);
 							return;
 						}
+
+						//remove specific value from key in storage
+						var tempData = JSON.parse(localStorage.getItem(key));
+						for (var i = tempData.length - 1; i >= 0; i--) {
+							if( JSON.stringify(value) == JSON.stringify(tempData[i]) ){
+								tempData.splice(i, 1);
+								localStorage.setItem(key, tempData);
+								callback(JSON.parse(localStorage.getItem(key)), null);
+								return;
+							}
+						}
+						callback(null, {"message": "no item has been removed","code": ""});
+						return;
+					}else{
+						console.warn("clear local storage while there is locked items not implemented");
+						callback(null, null);
 					}
-					callback(null, {"message": "no item has been removed","code": ""});
-					return;
 				}catch(e){
 					$rootScope.currentScope.Toast.show("Error!","There was an error in clearing local storage", Toast.LONG, Toast.ERROR);
 	                console.error("clear: ", e);
@@ -195,12 +200,18 @@
 			getWorkspaceData: function(stringType, callback){
 				try{
 					var dataToRetrieve = localStorage["com.intel.userdata"];
-					if(stringType == true)
-						callback(dataToRetrieve, null);
+					if(stringType == true){
+						strDecompress(dataToRetrieve, function(stepsComp){
+							callback(stepsComp, null);
+						});
+					}
 					else{
 						if(dataToRetrieve){
-							dataToRetrieve = JSON.parse(dataToRetrieve);
-							callback(dataToRetrieve, null);
+							strDecompress(dataToRetrieve, function(stepsComp){
+								dataToRetrieve = JSON.parse(stepsComp);
+								callback(dataToRetrieve, null);
+							});
+
 						}else{
 							var data = {
 								"Steps": null,
@@ -243,8 +254,11 @@
 								"CurrentUser": currentUser
 							};
 						}
-						localStorage.setItem("com.intel.userdata", JSON.stringify(data));
-						callback(true, null);
+						console.log(data);
+						strCompress(JSON.stringify(data), function(stepsComp){
+							localStorage.setItem("com.intel.userdata", stepsComp);
+							callback(true, null);
+						});
 					});
 
 				}catch(e){
